Rename React components
