+++ app/javascript/components/example/IterationsSummaryTable.jsx
