+++ app/javascript/components/maintaining/IterationsSummaryTable.jsx
