+++ app/javascript/components/notifications/Icon.jsx
